+++ tests/Support/ApiMock.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Swagger\Tests\Support;
+namespace Yiisoft\Yii\Swagger\Tests\Support;
 
 use Psr\Http\Message\ResponseInterface;
 use Psr\Http\Message\ServerRequestInterface;