+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Enh #112: Bump `yiisoft/assets` to version `5.1.1`, refactor `SwaggerUiAsset` (@vjik)
 - Chg #113: Change PHP version constraint to `8.1 - 8.4` (@rustamwin)
 - Chg #115: Bump `zircote/swagger-php` to version `^5.0`. Drop version `^4.1` of `swagger-api/swagger-ui`  (@rustamwin)
+- Chg #116: Change project namespace to `Yiisoft\Yii\Swagger` (@rustamwin)
 
 ## 2.2.0 January 27, 2025